add sending interface and replace receiving
